Add pegged variables to paramaters
paramaters accept a 'vars' keyword that take var_name -> list
key/values. Lists must be the same length as paramater value,  these
variables are added to extra vars based on the paramater value. This
allows for things like
```yaml
    - !paramater
      name: "num_instances"
      values: [4, 8, 16, 64]
      vars:
        provision_opt_3: ["60G", "124G", "252G", "1020G"]
```
where you want to increase some variable based on a paramater,  but
DON'T want it to be apart of the paramater product.